ali: update classes

+++ src/main/java/com/mftplus/demo/model/service/PaymentService.java
@@ -36,7 +36,9 @@ public class PaymentService implements Service<Payment, Long> {
     @Override
     public void remove(Long id) {
         Payment payment = entityManager.find(Payment.class, id);
-        entityManager.remove(payment);
+        if (payment != null) {
+            entityManager.remove(payment);
+        }
     }
 
     @Transactional
@@ -53,9 +55,30 @@ public class PaymentService implements Service<Payment, Long> {
     }
 
     @Transactional
-    public List<Payment> findByDate(LocalDate date) {
-        Query query = entityManager.createQuery("select p from paymentEntity p where p.date = :date", Payment.class);
-        query.setParameter("date", date);
+    public Payment findByDocNumber(Long docNumber) {
+        Query query = entityManager.createQuery("select p from paymentEntity p where p.docNumber = :docNumber", Payment.class);
+        query.setParameter("docNumber", docNumber);
+        return (Payment) query.getSingleResult();
+    }
+
+    @Transactional
+    public List<Payment> findByPaymentDate(LocalDate paymentDate) {
+        Query query = entityManager.createQuery("select p from paymentEntity p where p.date = :paymentDate", Payment.class);
+        query.setParameter("paymentDate", paymentDate);
+        return query.getResultList();
+    }
+
+    @Transactional
+    public List<Payment> findByPaymentMethod(String paymentMethod) {
+        Query query = entityManager.createQuery("select p from paymentEntity p where p.paymentMethod = :paymentMethod", Payment.class);
+        query.setParameter("paymentMethod", paymentMethod);
+        return query.getResultList(); // اصلاح شده
+    }
+
+    @Transactional
+    public List<Payment> findByTransactionId(Long transactionId) {
+        Query query = entityManager.createQuery("select p from paymentEntity p where p.transaction.id = :transactionId", Payment.class);
+        query.setParameter("transactionId", transactionId);
         return query.getResultList();
     }
 }
